Implementando exceções customizadas de atualizar e deletar

+++ src/main/java/com/devsipriano/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.devsipriano.dscommerce.services;
 import com.devsipriano.dscommerce.dto.ProductDTO;
 import com.devsipriano.dscommerce.entities.Product;
 import com.devsipriano.dscommerce.repositories.ProductRepository;
+import com.devsipriano.dscommerce.services.exceptions.DatabaseException;
 import com.devsipriano.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -42,15 +46,29 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id); //Não vai no banco de dados, mas instancia com id e fica monitorado pela JPA, diferente do new
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id); //Não vai no banco de dados, mas instancia com id e fica monitorado pela JPA, diferente do new
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso Não encontrado");
+        }
     }
 
-    @Transactional()
+    //Só executa o Transactional se o método se estiver no contexto de outra transação
+    //É pra captura o erro certo nesse contexto do h2
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
